+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ActFramework Change Log
 
 **1.5.2 TBD**
+* Handle login redirect gracefully #412
+* `NPE` while returning an `InputStream` in an action handler #410
 * Mal behavior of app running in dev mode with `.version` file #409
 
 **1.5.1 15/Nov/2017**